requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
pydantic==1.10.7
python-dotenv==1.0.0
asyncpg==0.27.0
alembic==1.10.2
pytest==7.3.1
httpx==0.24.0
python-jose==3.3.0
passlib==1.7.4